feat : articleFilter 함수
